// server/@types/shared/models/SessionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AppointmentSummaryDto } from './AppointmentSummaryDto';
export type SessionDto = {
    projectName: string;
    projectCode: string;
    projectLocation: string;
    date: string;
    startTime: string;
    endTime: string;
    appointmentSummaries: Array<AppointmentSummaryDto>;
};